+++ lib/commercetools-api/src/Models/ShippingMethod/ShippingMethod.php
@@ -12,6 +12,7 @@ use Commercetools\Api\Models\Common\BaseResource;
 use Commercetools\Api\Models\Common\CreatedBy;
 use Commercetools\Api\Models\Common\LastModifiedBy;
 use Commercetools\Api\Models\Common\LocalizedString;
+use Commercetools\Api\Models\Store\StoreKeyReferenceCollection;
 use Commercetools\Api\Models\TaxCategory\TaxCategoryReference;
 use Commercetools\Api\Models\Type\CustomFields;
 use Commercetools\Base\DateTimeImmutableCollection;
@@ -33,6 +34,7 @@ interface ShippingMethod extends BaseResource
     public const FIELD_IS_DEFAULT = 'isDefault';
     public const FIELD_PREDICATE = 'predicate';
     public const FIELD_CUSTOM = 'custom';
+    public const FIELD_STORES = 'stores';
 
     /**
      * <p>Unique identifier of the ShippingMethod.</p>
@@ -172,6 +174,17 @@ interface ShippingMethod extends BaseResource
     public function getCustom();
 
     /**
+     * <ul>
+     * <li>If a value exists, the Shipping Method applies to <a href="ctp:api:type:Cart">Carts</a> with a <a href="ctp:api:type:Store">Store</a> that matches any Store in this field.</li>
+     * <li>If empty, the Shipping Method applies to all <a href="ctp:api:type:Cart">Carts</a>, irrespective of a Store.</li>
+     * </ul>
+     *
+
+     * @return null|StoreKeyReferenceCollection
+     */
+    public function getStores();
+
+    /**
      * @param ?string $id
      */
     public function setId(?string $id): void;
@@ -255,4 +268,9 @@ interface ShippingMethod extends BaseResource
      * @param ?CustomFields $custom
      */
     public function setCustom(?CustomFields $custom): void;
+
+    /**
+     * @param ?StoreKeyReferenceCollection $stores
+     */
+    public function setStores(?StoreKeyReferenceCollection $stores): void;
 }