+++ lib/commercetools-api/src/Models/ShippingMethod/ShippingMethodAddStoreAction.php
@@ -1,0 +1,31 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * This file has been auto generated
+ * Do not change it.
+ */
+
+namespace Commercetools\Api\Models\ShippingMethod;
+
+use Commercetools\Api\Models\Store\StoreResourceIdentifier;
+use Commercetools\Base\DateTimeImmutableCollection;
+use Commercetools\Base\JsonObject;
+
+interface ShippingMethodAddStoreAction extends ShippingMethodUpdateAction
+{
+    public const FIELD_STORE = 'store';
+
+    /**
+     * <p>ResourceIdentifier of the Store to add.</p>
+     *
+
+     * @return null|StoreResourceIdentifier
+     */
+    public function getStore();
+
+    /**
+     * @param ?StoreResourceIdentifier $store
+     */
+    public function setStore(?StoreResourceIdentifier $store): void;
+}