+++ lib/commercetools-api/src/Models/ShippingMethod/ShippingMethodDraftBuilder.php
@@ -10,6 +10,7 @@ namespace Commercetools\Api\Models\ShippingMethod;
 
 use Commercetools\Api\Models\Common\LocalizedString;
 use Commercetools\Api\Models\Common\LocalizedStringBuilder;
+use Commercetools\Api\Models\Store\StoreResourceIdentifierCollection;
 use Commercetools\Api\Models\TaxCategory\TaxCategoryResourceIdentifier;
 use Commercetools\Api\Models\TaxCategory\TaxCategoryResourceIdentifierBuilder;
 use Commercetools\Api\Models\Type\CustomFieldsDraft;
@@ -91,6 +92,12 @@ final class ShippingMethodDraftBuilder implements Builder
      * @var null|CustomFieldsDraft|CustomFieldsDraftBuilder
      */
     private $custom;
+
+    /**
+
+     * @var ?StoreResourceIdentifierCollection
+     */
+    private $stores;
 
     /**
      * <p>User-defined unique identifier for the ShippingMethod.</p>
@@ -211,6 +218,21 @@ final class ShippingMethodDraftBuilder implements Builder
     public function getCustom()
     {
         return $this->custom instanceof CustomFieldsDraftBuilder ? $this->custom->build() : $this->custom;
+    }
+
+    /**
+     * <ul>
+     * <li>If defined and not empty, the Shipping Method applies to <a href="ctp:api:type:Cart">Carts</a> with a <a href="ctp:api:type:Store">Store</a> that matches any Store in this field.</li>
+     * <li>If not defined or empty, the Shipping Method applies to all Carts, irrespective of a Store.</li>
+     * </ul>
+     * <p>If the number of referenced Stores exceeds the <a href="/api/limits#shipping-methods">Stores per Shipping Method limit</a>, an <a href="ctp:api:type:InvalidOperationError">InvalidOperation</a> error is returned.</p>
+     *
+
+     * @return null|StoreResourceIdentifierCollection
+     */
+    public function getStores()
+    {
+        return $this->stores;
     }
 
     /**
@@ -335,6 +357,17 @@ final class ShippingMethodDraftBuilder implements Builder
     }
 
     /**
+     * @param ?StoreResourceIdentifierCollection $stores
+     * @return $this
+     */
+    public function withStores(?StoreResourceIdentifierCollection $stores)
+    {
+        $this->stores = $stores;
+
+        return $this;
+    }
+
+    /**
      * @deprecated use withLocalizedName() instead
      * @return $this
      */
@@ -391,7 +424,8 @@ final class ShippingMethodDraftBuilder implements Builder
             $this->active,
             $this->isDefault,
             $this->predicate,
-            $this->custom instanceof CustomFieldsDraftBuilder ? $this->custom->build() : $this->custom
+            $this->custom instanceof CustomFieldsDraftBuilder ? $this->custom->build() : $this->custom,
+            $this->stores
         );
     }